requirements.txt
# Core automation framework dependencies
requests>=2.31.0
tqdm>=4.66.0
cryptography>=41.0.0
ipaddress>=1.0.23
configparser>=6.0.0

# Scheduling and cron support
croniter>=1.4.0
PyYAML>=6.0.1

# Network utilities
python-nmap>=0.7.1
ping3>=4.0.4
dnspython>=2.4.0

# Git operations
GitPython>=3.1.40

# Multiprocessing and async
multiprocessing-logging>=0.3.4

# Enhanced IPAM client
urllib3>=2.0.7

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0

# Optional performance monitoring
psutil>=5.9.0